Make the operation class generic to be used for everyone. Get rid of mutable variables by using Either.
